requirements.txt
openai-whisper @ git+https://github.com/openai/whisper.git
openai==0.27.8
moviepy